Move Modal to ModalWindow component
